backend/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.2
Flask-Migrate==3.1.0
bcrypt==3.2.0
python-dotenv==1.0.0
openai==1.2.0
spacy==3.5.0
elasticsearch==8.6.2
textstat==0.7.3
pytest==7.0.1
